Fix imports after package move
