+++ src/fz/corpus/analyzer.py
